Move prices into portal
